Add Slide entity and configure DI for Slide services
Introduced the Slide entity into the Shop Management system.
Configured the Slide entity's properties and constraints in the
SlideMapping class. Implemented the SlideRepository class to
interact with Slide data in the database. Updated ShopContext
to include the Slide entity for management by Entity Framework
Core. Registered ISlideApplication and ISlideRepository services
in the dependency injection container.