Add constructor to CreateInventory; update form handler
A constructor was added to the CreateInventory class in the
CreateInventory.cs file. This constructor initializes the
Products property as a new list of ProductViewModel objects.

In the Create.cshtml file, the form's asp-page-handler
attribute was changed from "Define" to "Create". This change
modifies the behavior of the form submission process.