Add permissions to Role entity and update EF Core migration
Introduce a protected constructor to the Role class and modify the existing constructor to initialize the Permissions property. Update AccountContextModelSnapshot.cs to configure the RolePermissions table with necessary columns and relationships. Add a new migration class 20250317012643_PermissionAddedToRole to create the RolePermissions table in the database. Generate 20250317012643_PermissionAddedToRole.Designer.cs to reflect the entity model changes.